Refactor integrated GUI into tab modules with shared app state
